Join label rendering configuration commands and rename some commands
