+++ src/main/java/org/embl/mobie/viewer/command/ConfigureLabelRenderingCommand.java
@@ -2,6 +2,8 @@ package org.embl.mobie.viewer.command;
 
 import bdv.util.BdvHandle;
 import bdv.viewer.SourceAndConverter;
+import de.embl.cba.tables.color.CategoryColoringModel;
+import de.embl.cba.tables.color.ColoringModel;
 import net.imglib2.type.numeric.ARGBType;
 import org.embl.mobie.viewer.color.LabelConverter;
 import org.embl.mobie.viewer.source.LabelSource;
@@ -9,6 +11,7 @@ import org.embl.mobie.viewer.source.SourceHelper;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 import org.scijava.util.ColorRGB;
+import org.scijava.widget.Button;
 import sc.fiji.bdvpg.scijava.command.BdvPlaygroundActionCommand;
 
 import java.util.Arrays;
@@ -39,6 +42,28 @@ public class ConfigureLabelRenderingCommand implements BdvPlaygroundActionComman
 
 	@Parameter ( label = "Opacity of non-selected labels" )
 	public double opacity = 0.15;
+
+	@Parameter( label = "Increment random color seed [ Ctrl L ]", callback = "incrementRandomColorSeed" )
+	Button button;
+
+	public static void incrementRandomColorSeed( SourceAndConverter[] sourceAndConverters )
+	{
+		for ( SourceAndConverter sourceAndConverter : sourceAndConverters )
+		{
+			if ( sourceAndConverter.getConverter() instanceof LabelConverter< ? > )
+			{
+				final LabelConverter< ? > converter = ( LabelConverter< ? > ) sourceAndConverter.getConverter();
+
+				final ColoringModel< ? > coloringModel = converter.getColoringModel().getWrappedColoringModel();
+				if ( coloringModel instanceof CategoryColoringModel )
+				{
+					int randomSeed = ( ( CategoryColoringModel< ? > ) coloringModel ).getRandomSeed();
+					( ( CategoryColoringModel<?> ) coloringModel ).setRandomSeed( ++randomSeed );
+				}
+			}
+		}
+	}
+
 
 	@Override
 	public void run()
